Restructure the window package
