+++ src/builtin_plugins/plugin/amulet/editor/_main.py
@@ -21,7 +21,7 @@ from plugin.tablericons import tablericons
 from plugin.amulet.level import get_main_level, set_main_level
 
 from plugin.amulet.editor import __path__ as editor_plugin_path
-from plugin.amulet.editor.window._main import (
+from plugin.amulet.editor.window._main_window import (
     init_main_window,
     get_main_window,
     destroy_main_window,
