+++ src/builtin_plugins/plugin/amulet/editor/window/__init__.py
@@ -1,2 +1,2 @@
-from ._main import ButtonProxy
+from ._toolbar import ButtonProxy
 from ._tab_engine import TabWidget
