--- src/builtin_plugins/plugin/amulet/editor/window/_child/__init__.py
@@ -1,1 +1,0 @@
-from .child_window import AmuletChildWindow, create_sub_window, sub_windows
