--- src/builtin_plugins/plugin/amulet/editor/window/_main/__init__.py
@@ -1,2 +1,0 @@
-from .toolbar import ButtonProxy
-from ._main_window import AmuletMainWindow, init_main_window, get_main_window, destroy_main_window
